Use jQuery to sequentially fade items in
Previously, a CSS property was used to delay animations (was hardcoded into the element :fearful:). To fix this, add a "hidden" CSS class that hides the elements that are to be animated. Then use jQuery to iterate over DOMs with this class, add the animation class and remove the hidden class. This implementation waits one second between elements.